requirements.txt
confluent-kafka[avro]
python-dateutil
bugsnag